refactor(kubernetes-provider): address review feedback on exposure resolution
- ensureService: only preserve allocated nodePorts for NodePort/LoadBalancer
  Services (copying a nodePort onto a ClusterIP Service would be rejected), and
  preserve dual-stack immutable fields (clusterIPs, ipFamilies, ipFamilyPolicy).
- ResolveExposure: return ErrDynamicClientRequired when a GatewayClassName is set
  but the dynamic client is nil, and emit a warning on Gateway failure before
  falling back to LoadBalancer/NodePort instead of silently swallowing the error.
- Remove the now-unused ErrGatewayClassNotFound sentinel.
- Fix Kind/vCluster exposure comments to reflect that the LoadBalancer tier is
  skipped (Gateway → NodePort), and document the Gateway fallback in the docs.
- Move the buildValuesFiles test into provisioner_test.go.
