+++ pkg/svc/provisioner/cluster/vcluster/provisioner_test.go
@@ -3,6 +3,7 @@ package vclusterprovisioner_test
 import (
 	"context"
 	"errors"
+	"os"
 	"testing"
 
 	"github.com/devantler-tech/ksail/v7/pkg/svc/provider"
@@ -311,4 +312,42 @@ func TestExists_DefaultName(t *testing.T) {
 	require.NoError(t, err)
 	assert.True(t, exists)
 	mockProvider.AssertExpectations(t)
+}
+
+func TestBuildValuesFiles_ExtraSAN(t *testing.T) {
+	t.Parallel()
+
+	t.Run("adds proxy extraSANs when address provided", func(t *testing.T) {
+		t.Parallel()
+
+		files, cleanup, err := vclusterprovisioner.BuildValuesFilesForTest("", false, "5.6.7.8")
+		require.NoError(t, err)
+
+		defer cleanup()
+
+		require.NotEmpty(t, files)
+
+		content, err := os.ReadFile(files[0])
+		require.NoError(t, err)
+
+		assert.Contains(t, string(content), "proxy:")
+		assert.Contains(t, string(content), "extraSANs:")
+		assert.Contains(t, string(content), "5.6.7.8")
+	})
+
+	t.Run("omits proxy extraSANs when address empty", func(t *testing.T) {
+		t.Parallel()
+
+		files, cleanup, err := vclusterprovisioner.BuildValuesFilesForTest("", false, "")
+		require.NoError(t, err)
+
+		defer cleanup()
+
+		require.NotEmpty(t, files)
+
+		content, err := os.ReadFile(files[0])
+		require.NoError(t, err)
+
+		assert.NotContains(t, string(content), "extraSANs:")
+	})
 }
